CFrame Byte Optimization
Save up to 5 bytes per CFrame with this optimization. Roblox internally uses these values when creating a rbxl binary.  Stores a table of common rotations and stores in a list then sends only one byte instead of the quaternion when it hits. Only problem is if it misses it adds an extra byte.