Update 1.4.0
This update primarily focused on making it possible to change settings while writing and reading, without having to clone the module to do so.

To facilitate this. All parameters past the 2nd have been merged into one parameter that takes a table.

You can read the updated api in the readme file.